# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Spinlab Deal Desk API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /deals:
    get:
      operationId: listDeals
      tags: [deals]
      summary: List all deals
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: priority
          in: query
          schema:
            type: string
        - name: brokerId
          in: query
          schema:
            type: ["integer", "null"]
        - name: search
          in: query
          schema:
            type: string
        - name: city
          in: query
          schema:
            type: string
        - name: state
          in: query
          schema:
            type: string
        - name: overdueOnly
          in: query
          schema:
            type: boolean
        - name: followUpToday
          in: query
          schema:
            type: boolean
        - name: hotOnly
          in: query
          schema:
            type: boolean
        - name: deadOnly
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of deals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Deal"
    post:
      operationId: createDeal
      tags: [deals]
      summary: Create a new deal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DealInput"
      responses:
        "201":
          description: Created deal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deal"
  /deals/{id}:
    get:
      operationId: getDeal
      tags: [deals]
      summary: Get a deal by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deal details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deal"
        "404":
          description: Not found
    patch:
      operationId: updateDeal
      tags: [deals]
      summary: Update a deal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DealUpdate"
      responses:
        "200":
          description: Updated deal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deal"
    delete:
      operationId: deleteDeal
      tags: [deals]
      summary: Delete a deal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /deals/{id}/red-flags:
    get:
      operationId: getDealRedFlags
      tags: [deals]
      summary: Get red flags for a deal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Red flag checklist
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RedFlag"
    put:
      operationId: updateDealRedFlags
      tags: [deals]
      summary: Update red flags for a deal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RedFlagUpdate"
      responses:
        "200":
          description: Updated red flags
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RedFlag"

  # ─── BROKERS ──────────────────────────────────────────────────────────────
...
  /notes:
    get:
      operationId: listNotes
      tags: [notes]
      summary: List notes
      parameters:
        - name: linkedType
          in: query
          schema:
            type: string
        - name: linkedId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Note"
    post:
      operationId: createNote
      tags: [notes]
      summary: Create a new note
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteInput"
      responses:
        "201":
          description: Created note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
  /notes/{id}:
    patch:
      operationId: updateNote
      tags: [notes]
      summary: Update a note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteUpdate"
      responses:
        "200":
          description: Updated note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
    delete:
      operationId: deleteNote
      tags: [notes]
      summary: Delete a note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── DOCUMENTS ────────────────────────────────────────────────────────────
  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List documents for a deal
      parameters:
        - name: dealId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
  /documents/{id}:
    patch:
      operationId: updateDocument
      tags: [documents]
      summary: Update a document status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentUpdate"
      responses:
        "200":
          description: Updated document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
components:
  schemas:
...
    Deal:
      type: object
      required:
        - id
        - dealName
        - status
        - priority
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
        dealName:
          type: string
        businessName:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        state:
          type: ["string", "null"]
        assetType:
          type: ["string", "null"]
        source:
          type: ["string", "null"]
        brokerId:
          type: ["integer", "null"]
        brokerName:
          type: ["string", "null"]
        sellerName:
          type: ["string", "null"]
        askingPrice:
          type: ["number", "null"]
        grossRevenue:
          type: ["number", "null"]
        netIncome:
          type: ["number", "null"]
        adjustedNetIncome:
          type: ["number", "null"]
        monthlyRent:
          type: ["number", "null"]
        annualRent:
          type: ["number", "null"]
        leaseYearsRemaining:
          type: ["number", "null"]
        renewalOptions:
          type: ["string", "null"]
        squareFootage:
          type: ["number", "null"]
        rentPerSqFt:
          type: ["number", "null"]
        numWashers:
          type: ["integer", "null"]
        numDryers:
          type: ["integer", "null"]
        machineBrand:
          type: ["string", "null"]
        avgMachineAge:
          type: ["number", "null"]
        cardOrCoin:
          type: ["string", "null"]
        washAndFold:
          type: ["boolean", "null"]
        pickupDelivery:
          type: ["boolean", "null"]
        commercialAccounts:
          type: ["boolean", "null"]
        hoursOfOperation:
          type: ["string", "null"]
        staffCount:
          type: ["integer", "null"]
        ownerOperated:
          type: ["boolean", "null"]
        status:
          type: string
        priority:
          type: string
        nextAction:
          type: ["string", "null"]
        nextActionDueDate:
          type: ["string", "null"]
        lastContactedDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        askingMultiple:
          type: ["number", "null"]
        rentAsPercentGross:
          type: ["number", "null"]
        redFlagScore:
          type: integer
        redFlagLevel:
          type: string
        dealScore:
          type: integer
        dealQuality:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
    DealInput:
      type: object
      required: [dealName, status, priority]
      properties:
        dealName:
          type: string
        businessName:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        assetType:
          type: string
        source:
          type: string
        brokerId:
          type: ["integer", "null"]
        sellerName:
          type: string
        askingPrice:
          type: ["number", "null"]
        grossRevenue:
          type: ["number", "null"]
        netIncome:
          type: ["number", "null"]
        adjustedNetIncome:
          type: ["number", "null"]
        monthlyRent:
          type: ["number", "null"]
        leaseYearsRemaining:
          type: ["number", "null"]
        renewalOptions:
          type: string
        squareFootage:
          type: ["number", "null"]
        numWashers:
          type: ["integer", "null"]
        numDryers:
          type: ["integer", "null"]
        machineBrand:
          type: string
        avgMachineAge:
          type: ["number", "null"]
        cardOrCoin:
          type: string
        washAndFold:
          type: boolean
        pickupDelivery:
          type: boolean
        commercialAccounts:
          type: boolean
        hoursOfOperation:
          type: string
        staffCount:
          type: ["integer", "null"]
        ownerOperated:
          type: boolean
        status:
          type: string
        priority:
          type: string
        nextAction:
          type: string
        nextActionDueDate:
          type: ["string", "null"]
        lastContactedDate:
          type: ["string", "null"]
        notes:
          type: string
    DealUpdate:
      type: object
      properties:
        dealName:
          type: string
        businessName:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        assetType:
          type: string
        source:
          type: string
        brokerId:
          type: ["integer", "null"]
        sellerName:
          type: string
        askingPrice:
          type: ["number", "null"]
        grossRevenue:
          type: ["number", "null"]
        netIncome:
          type: ["number", "null"]
        adjustedNetIncome:
          type: ["number", "null"]
        monthlyRent:
          type: ["number", "null"]
        leaseYearsRemaining:
          type: ["number", "null"]
        renewalOptions:
          type: string
        squareFootage:
          type: ["number", "null"]
        numWashers:
          type: ["integer", "null"]
        numDryers:
          type: ["integer", "null"]
        machineBrand:
          type: string
        avgMachineAge:
          type: ["number", "null"]
        cardOrCoin:
          type: string
        washAndFold:
          type: boolean
        pickupDelivery:
          type: boolean
        commercialAccounts:
          type: boolean
        hoursOfOperation:
          type: string
        staffCount:
          type: ["integer", "null"]
        ownerOperated:
          type: boolean
        status:
          type: string
        priority:
          type: string
        nextAction:
          type: string
        nextActionDueDate:
          type: ["string", "null"]
        lastContactedDate:
          type: ["string", "null"]
        notes:
          type: string

    # ─── RED FLAGS ─────────────────────────────────────────────────────────
    RedFlag:
      type: object
      required: [id, dealId, flagKey, flagLabel, checked]
      properties:
        id:
          type: integer
        dealId:
          type: integer
        flagKey:
          type: string
        flagLabel:
          type: string
        checked:
          type: boolean
    RedFlagUpdate:
      type: object
      required: [flags]
      properties:
        flags:
          type: array
          items:
            type: object
            required: [flagKey, checked]
            properties:
              flagKey:
                type: string
              checked:
                type: boolean

    # ─── BROKER ────────────────────────────────────────────────────────────
...
    Note:
      type: object
      required:
        - id
        - linkedType
        - linkedId
        - noteText
        - noteType
        - createdAt
      properties:
        id:
          type: integer
        linkedType:
          type: string
        linkedId:
          type: integer
        linkedName:
          type: ["string", "null"]
        noteType:
          type: string
        noteText:
          type: string
        createdAt:
          type: string
    NoteInput:
      type: object
      required: [linkedType, linkedId, noteText, noteType]
      properties:
        linkedType:
          type: string
        linkedId:
          type: integer
        noteType:
          type: string
        noteText:
          type: string
    NoteUpdate:
      type: object
      properties:
        noteType:
          type: string
        noteText:
          type: string

    # ─── DOCUMENT ──────────────────────────────────────────────────────────
    Document:
      type: object
      required:
        - id
        - dealId
        - docKey
        - docLabel
        - status
        - updatedAt
      properties:
        id:
          type: integer
        dealId:
          type: integer
        docKey:
          type: string
        docLabel:
          type: string
        status:
          type: string
        updatedAt:
          type: string
    DocumentUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string