# lib/api-zod/src/generated/types/dealInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Spinlab Deal Desk API
 * OpenAPI spec version: 0.1.0
 */

export interface DealInput {
  dealName: string;
  businessName?: string;
  address?: string;
  city?: string;
  state?: string;
  assetType?: string;
  source?: string;
  /** @nullable */
  brokerId?: number | null;
  sellerName?: string;
  /** @nullable */
  askingPrice?: number | null;
  /** @nullable */
  grossRevenue?: number | null;
  /** @nullable */
  netIncome?: number | null;
  /** @nullable */
  adjustedNetIncome?: number | null;
  /** @nullable */
  monthlyRent?: number | null;
  /** @nullable */
  leaseYearsRemaining?: number | null;
  renewalOptions?: string;
  /** @nullable */
  squareFootage?: number | null;
  /** @nullable */
  numWashers?: number | null;
  /** @nullable */
  numDryers?: number | null;
  machineBrand?: string;
  /** @nullable */
  avgMachineAge?: number | null;
  cardOrCoin?: string;
  washAndFold?: boolean;
  pickupDelivery?: boolean;
  commercialAccounts?: boolean;
  hoursOfOperation?: string;
  /** @nullable */
  staffCount?: number | null;
  ownerOperated?: boolean;
  status: string;
  priority: string;
  nextAction?: string;
  /** @nullable */
  nextActionDueDate?: string | null;
  /** @nullable */
  lastContactedDate?: string | null;
  notes?: string;
}