# lib/api-zod/src/generated/types/document.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Spinlab Deal Desk API
 * OpenAPI spec version: 0.1.0
 */

export interface Document {
  id: number;
  dealId: number;
  docKey: string;
  docLabel: string;
  status: string;
  updatedAt: string;
}