# lib/api-zod/src/generated/types/noteUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Spinlab Deal Desk API
 * OpenAPI spec version: 0.1.0
 */

export interface NoteUpdate {
  noteType?: string;
  noteText?: string;
}