# lib/api-zod/src/generated/types/redFlag.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Spinlab Deal Desk API
 * OpenAPI spec version: 0.1.0
 */

export interface RedFlag {
  id: number;
  dealId: number;
  flagKey: string;
  flagLabel: string;
  checked: boolean;
}