# lib/api-zod/src/generated/types/redFlagUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Spinlab Deal Desk API
 * OpenAPI spec version: 0.1.0
 */
import type { RedFlagUpdateFlagsItem } from './redFlagUpdateFlagsItem';

export interface RedFlagUpdate {
  flags: RedFlagUpdateFlagsItem[];
}